Move tensor files to graphapi subdirectory
